refactor(tests): split test_cli_setup.py into a focused cli_setup package (#443)
The 1033-LOC tests/test_cli_setup.py covered six unrelated t3 setup
concerns in one module. Split it verbatim into a tests/cli_setup/
package, one file per concern, each well under the 500-LOC
module-health limit. Test bodies are AST-identical to the original;
only import paths and per-concern helper placement changed.

Relates-to #443
